fix(insights): gate stickiness tooltip content behind flag; drop revenue analytics
- StickinessBarChart / StickinessLineChart: tooltip config was already
  flag-gated but renderTooltip content was not — add TrendsTooltip fallback
  and quillTooltipEnabled guard consistently with every other chart
- RevenueAnalyticsChart: revert — out of scope for this PR
